Remove pattern from pattern match
I copied this code and while I think it doesn't matter here I found that it was doing a pattern-match on the library path. So my library had a "-" in it and thus it didn't match, kept adding in the path. This uses some optional arguments to `find` that forces it to a literal match. Figured if someone else tries this they might also get confused :)